pyproject.toml
[project]
name = "daily-ai-info"
version = "0.1.0"
description = "Daily AI information crawler and summarizer"
authors = [{name = "Your Name", email = "your.email@example.com"}]
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "google-generativeai>=0.3.0",
    "apscheduler>=3.10.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "lxml>=4.9.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"